GH-157: fixes, button move
