+++ components/01-atoms/buttons/button.twig.test.js
@@ -1,8 +1,8 @@
 import { join } from 'path';
 import { render, Twig } from 'twig-testing-library';
 
-import loadYaml from '../../../../util/loadYaml';
-import { setupTwig } from '../../../../.storybook/setupTwig';
+import loadYaml from '../../../util/loadYaml';
+import { setupTwig } from '../../../.storybook/setupTwig';
 
 setupTwig(Twig);
