+++ package.json
@@ -78,10 +78,10 @@
   },
   "devDependencies": {
     "@babel/preset-react": "^7.12.5",
-    "@storybook/addon-a11y": "6.1.18",
+    "@storybook/addon-a11y": "^6.1.18",
     "@storybook/addon-essentials": "^6.1.18",
     "@storybook/addon-links": "^6.1.18",
-    "@storybook/addons": "6.1.18",
+    "@storybook/addons": "^6.1.18",
     "@storybook/html": "^6.1.18",
     "@storybook/storybook-deployer": "^2.8.7",
     "@testing-library/react": "^11.1.2",
